requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
python-dotenv==1.0.1
requests==2.31.0
openai-whisper==20231117
pydub==0.25.1
